Authorization at controller endpoints

+++ Vastra.API/Vastra.API/Controllers/UsersController.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.JsonPatch;
 using Microsoft.AspNetCore.Mvc;
 using System.Text.Json;
@@ -22,6 +23,7 @@ namespace Vastra.API.Controllers
             _mapper = mapper;
         }
         [HttpGet]
+        [Authorize(Policy = "MustBeAdmin")]
         public async Task<ActionResult<IEnumerable<UserDto>>> GetUsers(int roleId,
             string? name, string? searchQuery, int pageNumber = 1, int pageSize = 10)
         {
@@ -35,6 +37,7 @@ namespace Vastra.API.Controllers
         }
         [HttpHead]
         [HttpGet("{userId}", Name = "GetUser")]
+        [Authorize]
         public async Task<ActionResult<UserDto>> GetUser(int roleId, int userId, bool includeAddresses = false, bool includeOrders = false)
         {
             if(!await _vastraRepository.RoleExistsAsync(roleId))
@@ -42,6 +45,19 @@ namespace Vastra.API.Controllers
                 return NotFound();
             }
             var userEntity = await _vastraRepository.GetUserByRoleAsync(roleId, userId, includeAddresses, includeOrders);
+            if(userEntity == null)
+            {
+                return NotFound();
+            }
+            var claimedPhoneNumber = User.Claims.FirstOrDefault(c => c.Type.Equals("phone"));
+            if(claimedPhoneNumber == null)
+            {
+                return Forbid();
+            }
+            if (!claimedPhoneNumber.Value.Equals(userEntity.PhoneNumber))
+            {
+                return Forbid();
+            }
             if(includeAddresses && includeOrders)
             {
                 return Ok(_mapper.Map<UserWithAddressesAndOrdersDto>(userEntity));
@@ -62,6 +78,13 @@ namespace Vastra.API.Controllers
             if(!await _vastraRepository.RoleExistsAsync(roleId))
             {
                 return NotFound();
+            }
+            var role = await _vastraRepository.GetRoleAsync(roleId);
+            if(role.RoleName == "Admin") {
+                if (!User.IsInRole("Admin"))
+                {
+                    return Forbid();
+                }
             }
             var finalUser = _mapper.Map<Entities.User>(user);
 
@@ -99,6 +122,15 @@ namespace Vastra.API.Controllers
             {
                 return NotFound();
             }
+            var claimedPhoneNumber = User.Claims.FirstOrDefault(c => c.Type.Equals("phone"));
+            if (claimedPhoneNumber == null)
+            {
+                return Forbid();
+            }
+            if (!claimedPhoneNumber.Value.Equals(userEntity.PhoneNumber))
+            {
+                return Forbid();
+            }
             _mapper.Map(user, userEntity);
             //update Modified Time of product
             userEntity.DateModified = DateTime.Now;
@@ -117,6 +149,15 @@ namespace Vastra.API.Controllers
             if (userEntity == null)
             {
                 return NotFound();
+            }
+            var claimedPhoneNumber = User.Claims.FirstOrDefault(c => c.Type.Equals("phone"));
+            if (claimedPhoneNumber == null)
+            {
+                return Forbid();
+            }
+            if (!claimedPhoneNumber.Value.Equals(userEntity.PhoneNumber))
+            {
+                return Forbid();
             }
             var userToPatch = _mapper.Map<UserForUpdateDto>(userEntity);
             patchDocument.ApplyTo(userToPatch, ModelState);
@@ -146,6 +187,15 @@ namespace Vastra.API.Controllers
             if (userToDelete == null)
             {
                 return NotFound();
+            }
+            var claimedPhoneNumber = User.Claims.FirstOrDefault(c => c.Type.Equals("phone"));
+            if (claimedPhoneNumber == null)
+            {
+                return Forbid();
+            }
+            if (!claimedPhoneNumber.Value.Equals(userToDelete.PhoneNumber))
+            {
+                return Forbid();
             }
             _vastraRepository.DeleteUser(userToDelete);
             await _vastraRepository.SaveChangesAsync();
